+++ lib/zip_new.c
@@ -69,6 +69,7 @@ _zip_new(zip_error_t *error)
     za->nopen_source = za->nopen_source_alloc = 0;
     za->open_source = NULL;
     za->tempdir = NULL;
+    za->progress_callback = NULL;
     
     return za;
 }