+++ lib/zip_open.c
@@ -177,6 +177,14 @@ zip_archive_set_tempdir(zip_t *za, const char *tempdir)
     return 0;
 }
 
+
+ZIP_EXTERN void
+zip_register_progress_callback(zip_t *za, zip_progress_callback_t progress_callback)
+{
+    za->progress_callback = progress_callback;
+}
+
+
 zip_t *
 _zip_open(zip_source_t *src, unsigned int flags, zip_error_t *error)
 {