+++ lib/zipint.h
@@ -200,6 +200,7 @@ struct zip {
     zip_hash_t *names;			/* hash table for name lookup */
 
     char *tempdir;                      /* custom temp dir (needed e.g. for OS X sandboxing) */
+    zip_progress_callback_t progress_callback; /* progress callback for zip_close() */
 };
 
 /* file in zip archive, part of API */