+++ NEWS.md
@@ -6,6 +6,7 @@
 * Fix seeking in zip_source_file if start > 0.
 * Add zip_fseek() for seeking in uncompressed data.
 * Add zip_ftell() for telling position in uncompressed data.
+* Add zip_register_progress_callback() for UI updates during zip_close()
 
 1.1.3 [2016-05-28]
 ==================